+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.0.4 (11-16-2014)
+
+- Added support for creating new application from tiapp.xml
+- Added support for creating session from session id
+- Added support for retrieving cloud environments
+- Added support for both production and development logins
+- User.find will now use current user if not user id passed
+
 # 1.0.1 (09-29-2014)
 
 - Added support for creating an ACS application